+++ Rubrik/Public/New-RubrikLDAP.ps1
@@ -15,8 +15,21 @@ function New-RubrikLDAP
       https://rubrik.gitbook.io/rubrik-sdk-for-powershell/command-documentation/reference/new-rubrikldap
             
       .EXAMPLE
-      New-RubrikLDAP -Name "Test LDAP Settings" -baseDN "DC=domain,DC=local" -authServers "192.168.1.8"
+      New-RubrikLDAP -Name "Test LDAP Settings" -baseDn "DC=domain,DC=local" -authServers "192.168.1.8"
+
       This will create LDAP settings on the Rubrik cluster defined by Connect-Rubrik function
+
+      .EXAMPLE
+      $credential = Get-Credential
+      New-RubrikLDAP -Name "rubrik.lab" -DynamicDNSName "ad1.test.lab" -baseDn "DC=rubrik,DC=lab" -BindCredential $Credential -Verbose
+
+      This will create LDAP settings using the credentials object provided as a parameter
+
+      .EXAMPLE
+      $SecPw = Read-Host -AsSecureString
+      New-RubrikLDAP -Name "rubrik.lab" -DynamicDNSName "ad1.test.lab" -baseDn "DC=rubrik,DC=lab" -BindUserName jaapjaap -BindUserPassword $SecPw -Verbose
+
+      This will create LDAP settings using the user name and password provided as parameters      
   #>
 
   [cmdletbinding(SupportsShouldProcess=$true,DefaultParametersetName='UserPassword')]
@@ -32,7 +45,7 @@ function New-RubrikLDAP
     [Parameter(Mandatory=$True)]
     [string]$DynamicDNSName,
     # The path to the directory where searches for users begin.
-    [string]$BaseDN,
+    [string]$baseDn,
     # An ordered list of authentication servers. Servers on this list have priority over servers discovered using dynamic DNS.
     [array]$AuthServers,
     # Bind username with permissions to connect to the LDAP server
@@ -61,8 +74,8 @@ function New-RubrikLDAP
     Test-RubrikConnection
 
     # Check to ensure that we have credentials for the LDAP server
-    $BindCredential = Test-RubrikLDAPCredential -BindUserName $BindUserName -BindUserPassword $BindUserPassword -Credential $BindCredential
-    
+    $BindCredential = Test-RubrikLDAPCredential -BindUserName $BindUserName -BindUserPassword $BindUserPassword -BindCredential $BindCredential
+
     # API data references the name of the function
     # For convenience, that name is saved here to $function
     $function = $MyInvocation.MyCommand.Name
@@ -88,9 +101,21 @@ function New-RubrikLDAP
     # See this PR for more information: https://github.com/rubrikinc/rubrik-sdk-for-powershell/pull/263
     Write-Verbose 'Passing $BindCredential username and password into the API request'
     $bodyHash = ConvertFrom-Json $body
-    $bodyHash.bindUserName = $BindCredential.UserName
-    $bodyHash.bindUserPassword = $BindCredential.GetNetworkCredential().Password
+    
+    if (-not $bodyHash.bindUserName) {
+      Add-Member -InputObject $bodyhash -MemberType NoteProperty -Name 'bindUserName' -Value $BindCredential.UserName
+    } else {
+      $bodyHash.bindUserName = $BindCredential.UserName
+    }
+
+    if (-not $bodyHash.bindUserPassword) {
+      Add-Member -InputObject $bodyhash -MemberType NoteProperty -Name 'bindUserPassword' -Value $BindCredential.GetNetworkCredential().Password
+    } else {
+      $bodyHash.bindUserPassword = $BindCredential.GetNetworkCredential().Password
+    }
+
     $body = ConvertTo-Json $bodyHash
+    Write-Verbose -Message "Updated Body with credential object = $($body -replace 'bindUserPassword":\s*"(.*?)"','bindUserPassword": "***"')"
     #endregion    
 
     $result = Submit-Request -uri $uri -header $Header -method $($resources.Method) -body $body